// src/app/generated/model/materialRequestCreateRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MaterialRequestLineItemRequest } from './materialRequestLineItemRequest';


export interface MaterialRequestCreateRequest { 
    /**
     * Identifier of the site (a Warehouse of type SITE) this request is for
     */
    siteWarehouseId: number;
    /**
     * Date the materials are needed by, if known
     */
    dateNeeded?: string;
    /**
     * Optional free-text notes about the request
     */
    notes?: string;
    /**
     * Line items being requested; at least one is required
     */
    lines: Array<MaterialRequestLineItemRequest>;
}