// src/app/generated/model/materialRequestLineItemRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MaterialRequestLineItemRequest { 
    /**
     * Identifier of the item being requested
     */
    itemId: number;
    /**
     * Quantity of the item being requested
     */
    quantityRequested: number;
    /**
     * Optional free-text notes about this line
     */
    notes?: string;
}